Fundamental shift in how this plugin is used. The key method is now called "assert_valid_request", and now raises an exception (RequestError) if the request is not valid.  The old method was called "validate_request", and it would set an error message in the flash and redirect.
This is a better approach, since it turns out that it's quite easy to capture
exceptions at the top level of the application and deal with them there. This
is how the exception_notifier plugin does things (and we now mesh nicely with
that plugin as well). 

It also makes fewer assumptions about your application, since it's not creating
public error messages and putting them in the flash hash with a presumed key.

It also makes the method call cleaner, since you don't have to remember to add
"or return" to the end of each call, which was pretty clumsy.

Also, following convention with other plugins, it no longer automatically loads
itself -- you have to "include ValidateRequest" in your application.rb.



git-svn-id: https://secure.westarete.com/svn/internal/plugins/validate_request/trunk@7708 da5c320b-a5ec-0310-b7f9-e4d854caa1e2